+++ src/py_mono_tools/backends/docker.py
@@ -6,7 +6,7 @@ import typing as t
 from pathlib import PosixPath
 
 from py_mono_tools.backends.interface import Backend
-from py_mono_tools.utils.config import cfg, logger
+from py_mono_tools.config import cfg, logger
 
 
 # pylint: disable=R0801
