+++ src/py_mono_tools/config.py
@@ -4,8 +4,8 @@ import os
 import pathlib
 import typing as t
 
+from py_mono_tools.cli_interface import CliMachineOutput
 from py_mono_tools.goals.interface import Deployer, Linter, Tester
-from py_mono_tools.utils.cli_interface import CliMachineOutput
 
 
 if t.TYPE_CHECKING:
